Fixed some compilation warnings.
git-svn-id: https://europa-pso.googlecode.com/svn/PLASMA/trunk@1292 23e1bc29-d954-0410-9766-47de34158ba0

+++ Resource/component/ResourceTransactionConstraint.cc
@@ -18,7 +18,7 @@ namespace Prototype
 					 const ConstraintEngineId& constraintEngine,
 					 const std::vector<ConstrainedVariableId>& variables)
     : Constraint(name, propagatorName, constraintEngine, variables){
-    check_error(variables.size() == ARG_COUNT);
+    check_error(variables.size() == (unsigned int) ARG_COUNT);
     //@todo add type checking of each variable in the constraint
   }
